Move posts.js from src/database/model to src/model
